Move scan options out of dialog
