+++ ui/v2.5/src/locales/en-GB.json
@@ -78,6 +78,8 @@
     "select_all": "Select All",
     "select_folders": "Select folders",
     "select_none": "Select None",
+    "selective_auto_tag": "Selective Auto Tag",
+    "selective_scan": "Selective Scan",
     "set_as_default": "Set as default",
     "set_back_image": "Back image…",
     "set_front_image": "Front image…",
