+++ ui/v2.5/src/locales/es-ES.json
@@ -79,6 +79,8 @@
     "select_all": "Seleccionar todo",
     "select_folders": "Seleccionar carpetas",
     "select_none": "Deseleccionar todo",
+    "selective_auto_tag": "Etiquetado automático selectivo",
+    "selective_scan": "Búsqueda selectiva",
     "set_as_default": "Establecer por defecto",
     "set_back_image": "Contraportada…",
     "set_front_image": "Portada…",
