+++ ui/v2.5/src/locales/it-IT.json
@@ -79,6 +79,8 @@
     "select_all": "Seleziona Tutto",
     "select_folders": "Seleziona cartelle",
     "select_none": "Deseleziona Tutto",
+    "selective_auto_tag": "Tag Automatico Selettivo",
+    "selective_scan": "Scansione Selettiva",
     "set_as_default": "Imposta come Predefinito",
     "set_back_image": "Immagine Retro…",
     "set_front_image": "Immagine Frontale…",
