+++ ui/v2.5/src/locales/pt-BR.json
@@ -64,6 +64,8 @@
     "search": "Buscar",
     "select_all": "Selecionar todos",
     "select_none": "Selecionar nenhum",
+    "selective_auto_tag": "Auto Tag seletivo",
+    "selective_scan": "Escaneamento seletivo",
     "set_as_default": "Aplicar como padrão",
     "set_back_image": "Imagem de fundo…",
     "set_front_image": "Imagem frontal…",
