+++ ui/v2.5/src/locales/sv-SE.json
@@ -68,6 +68,8 @@
     "select_all": "Välj alla",
     "select_folders": "Välj mappar",
     "select_none": "Välj inga",
+    "selective_auto_tag": "Selektiv Auto Tag",
+    "selective_scan": "Selektiv skanning",
     "set_as_default": "Välj som standard",
     "set_back_image": "Bakbild…",
     "set_front_image": "Frambild…",
