+++ ui/v2.5/src/locales/zh-CN.json
@@ -79,6 +79,8 @@
     "select_all": "选择所有",
     "select_folders": "选择目录",
     "select_none": "清除选择",
+    "selective_auto_tag": "选择性自动生成标签",
+    "selective_scan": "选择性扫描",
     "set_as_default": "设置为默认",
     "set_back_image": "设置背面图…",
     "set_front_image": "设置正面图…",
