+++ ui/v2.5/src/locales/zh-TW.json
@@ -79,6 +79,8 @@
     "select_all": "全選",
     "select_folders": "選擇資料夾",
     "select_none": "清除選擇",
+    "selective_auto_tag": "選擇性套用標籤",
+    "selective_scan": "選擇性掃描",
     "set_as_default": "設為預設",
     "set_back_image": "設定背面圖…",
     "set_front_image": "設定正面圖…",
